+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0-beta.15] - 2026-05-08
 
+### Changed
+
+- arr instances are now enabled by default when creating new instances
+- Now shows all arr instances even disabled ones
+
 ### Fixed
 
 - Multiple arr instances not showing on dashboard (UI bug only)